Added token verification and user feedback to deleteSummonerThunk

+++ src/redux/thunks/summonersThunks.ts
@@ -10,8 +10,6 @@ import { AppDispatch } from "../store/store";
 
 const warningIcon = "/icons/warning-ico.webp";
 const errorIcon = "/icons/error-ico.webp";
-const errorLoginText =
-  ': "-No se ha podido borrar al invocador. Inténtalo de nuevo."';
 
 export const loadSummonersThunk = () => async (dispatch: AppDispatch) => {
   try {
@@ -31,24 +29,43 @@ export const deleteSummonerThunk =
   (idToDelete: string, summonerName: string) =>
   async (dispatch: AppDispatch) => {
     const succesLoginText = `: "-Adiosito ${summonerName}."`;
+    const token = localStorage.getItem("token");
 
-    try {
-      await axios.delete(
-        `${process.env.REACT_APP_API_URL}summoners/${idToDelete}`
-      );
+    if (token) {
+      try {
+        dispatch(loadOnActionCreator());
+        await axios.delete(
+          `${process.env.REACT_APP_API_URL}summoners/${idToDelete}`
+        );
 
-      toast.success(CustomToast(warningIcon, succesLoginText), {
+        toast.success(CustomToast(warningIcon, succesLoginText), {
+          position: "bottom-center",
+          hideProgressBar: true,
+          progress: undefined,
+        });
+
+        dispatch(deleteSummonerActionCreator(idToDelete));
+        dispatch(loadOffActionCreator());
+      } catch (error) {
+        const errorLoginText =
+          ': "-No se ha podido borrar al invocador. Inténtalo de nuevo."';
+
+        toast.error(CustomToast(errorIcon, errorLoginText), {
+          position: "bottom-center",
+          hideProgressBar: true,
+          progress: undefined,
+        });
+        dispatch(loadOffActionCreator());
+      }
+    } else {
+      const warningText =
+        '"-Tienes que estar loggeado para borrar a un invocador"';
+
+      toast.warning(CustomToast(warningIcon, warningText), {
         position: "bottom-center",
         hideProgressBar: true,
         progress: undefined,
       });
-
-      dispatch(deleteSummonerActionCreator(idToDelete));
-    } catch (error) {
-      toast.error(CustomToast(errorIcon, errorLoginText), {
-        position: "bottom-center",
-        hideProgressBar: true,
-        progress: undefined,
-      });
+      dispatch(loadOffActionCreator());
     }
   };
